pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.27.1",
    "pydantic>=2.5.0",
    "uv>=0.8.10",
    "requests>=2.31.0",
    "pypdf>=4.2.0",
    "pdfminer.six>=20221105",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.2",
    "beautifulsoup4>=4.12.3",
    "markdownify>=0.13.1",
    "openai>=1.30.1",
    "qdrant-client>=1.15.1",
    "langchain-core>=0.3.74",
    "langchain>=0.3.27",
    "langchain-text-splitters>=0.3.9",
    "chromadb>=1.0.20",
]
